app/requirements.txt
fastapi
uvicorn
psycopg2-binary
requests
beautifulsoup4
pandas